examples/python-handler/requirements.txt
graphene==2.0.1
Flask==0.12.2
Flask-GraphQL==1.4.1
dbus-python
